requirements.txt
streamlitpandasnumpy
matplotlib
seaborn
scikit-learn
joblib
